Add files via upload
This calculator is based off of the Philosopher Richard Swineburne's book "The Incarnation of God Incarnate" which he used Bayesian probability to measure the likelihood of the supposed past event of the resurrection of Jesus of Nazareth in history. I built the calculator to run numerous trials and weigh how various inputs effect the final output. It also can function as a great tool for illuminating background beliefs based on the initial questions from different worldviews who may try to run the calculator.